Create SMA_EMA_Cross_Stratery.pine
Plotted SMA and EMA and achieved progress in printing the buy/sell signal. 
I tested the code with different time frames and the results differed depending on the coin and the time frame. best results were a win rate of 40 percent of all trades taken. 
Development is still in progress.